third_party/blink/web_tests/fast/dom/shadow/activeelement-should-be-shadowhost-expected.txt

CONSOLE WARNING: line 32: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Makes sure that document.activeElement returns a shadow host when a element in the correspoinding shadow tree is focused.


Focusing top/input0
PASS: top document.activeElement is top/shadow0
PASS: top/iframe1 document.activeElement is top/iframe1/doc1_body
PASS: top/iframe3 document.activeElement is top/iframe3/doc3_body
PASS: top/iframe4 document.activeElement is top/iframe4/doc4_body
PASS: top/iframe4/iframe4_1 document.activeElement is top/iframe4/iframe4_1/doc4_1_body

Focusing top/iframe1/input1
PASS: top document.activeElement is top/iframe1
PASS: top/iframe1 document.activeElement is top/iframe1/shadow1
PASS: top/iframe3 document.activeElement is top/iframe3/doc3_body
PASS: top/iframe4 document.activeElement is top/iframe4/doc4_body
PASS: top/iframe4/iframe4_1 document.activeElement is top/iframe4/iframe4_1/doc4_1_body

Focusing top/input2
PASS: top document.activeElement is top/shadow2
PASS: top/iframe1 document.activeElement is top/iframe1/doc1_body
PASS: top/iframe3 document.activeElement is top/iframe3/doc3_body
PASS: top/iframe4 document.activeElement is top/iframe4/doc4_body
PASS: top/iframe4/iframe4_1 document.activeElement is top/iframe4/iframe4_1/doc4_1_body

Focusing top/iframe3/input3
PASS: top document.activeElement is top/shadow3
PASS: top/iframe1 document.activeElement is top/iframe1/doc1_body
PASS: top/iframe3 document.activeElement is top/iframe3/input3
PASS: top/iframe4 document.activeElement is top/iframe4/doc4_body
PASS: top/iframe4/iframe4_1 document.activeElement is top/iframe4/iframe4_1/doc4_1_body

Focusing top/iframe4/iframe4_1/input4
PASS: top document.activeElement is top/iframe4
PASS: top/iframe1 document.activeElement is top/iframe1/doc1_body
PASS: top/iframe3 document.activeElement is top/iframe3/doc3_body
PASS: top/iframe4 document.activeElement is top/iframe4/iframe4_1
PASS: top/iframe4/iframe4_1 document.activeElement is top/iframe4/iframe4_1/input4
